+++ lib/Trace/TraceReader.cpp
@@ -162,9 +162,9 @@ ProcessTrace::ProcessTrace(InputBufferAllocator &Allocator,
                            std::string &&ModuleIdentifier,
                            uint32_t NumThreads,
                            uint64_t FinalProcessTime,
-                           std::vector<uint64_t> &&GVAddresses,
+                           std::vector<uintptr_t> &&GVAddresses,
                            std::vector<offset_uint> &&GVInitialData,
-                           std::vector<uint64_t> &&FAddresses,
+                           std::vector<uintptr_t> &&FAddresses,
                            std::vector<std::unique_ptr<ThreadTrace>> &&TTraces
                            )
 : BufferAllocator(Allocator),
@@ -199,9 +199,9 @@ ProcessTrace::readFrom(InputBufferAllocator &Allocator) {
   std::string ModuleIdentifier;
   uint32_t NumThreads;
   uint64_t FinalProcessTime;
-  std::vector<uint64_t> GlobalVariableAddresses;
+  std::vector<uintptr_t> GlobalVariableAddresses;
   std::vector<offset_uint> GlobalVariableInitialData;
-  std::vector<uint64_t> FunctionAddresses;
+  std::vector<uintptr_t> FunctionAddresses;
   std::vector<std::unique_ptr<ThreadTrace>> ThreadTraces;
 
   TraceReader >> ModuleIdentifier
